requirements.txt
# Enhanced Model Dependencies

# Core ML
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Deep Learning (optional, for BERT)
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0

# Sentence Embeddings
sentence-transformers>=2.2.0

# BM25 for hybrid search
rank-bm25>=0.2.2

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# LLM
groq>=0.4.0
python-dotenv>=1.0.0

# Reporting
reportlab>=4.0.0

# Utils
joblib>=1.3.0